requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-bootstrap5==24.2
django-pytest==0.2.0
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==25.3.0
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
